seperated methods from console
The console section isn't going to service everyone and the methods need split up so they serve use cases (that will help group development too).

This may need rejigged (currently this is done by URL rather than use case) but for now I'm happy to do it this way as it should make development easier
