+++ FantasyPremierLeague.pyproj
@@ -22,6 +22,12 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="FantasyPremierLeague.py" />
+    <Compile Include="gameweekSummary.py">
+      <SubType>Code</SubType>
+    </Compile>
+    <Compile Include="playerData.py">
+      <SubType>Code</SubType>
+    </Compile>
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Python Tools\Microsoft.PythonTools.targets" />
   <!-- Uncomment the CoreCompile target to enable the Build command in
